+++ packages/bitcoincash/src/signers/trezor.signer.ts
@@ -1,4 +1,3 @@
-import * as BitcoinCash from 'bitcoinjs-lib';
 import { isValidAddress } from 'bchaddrjs';
 import {
   Signer,
@@ -85,7 +84,7 @@ export class TrezorSigner extends Signer.TrezorProvider {
       };
     });
 
-    const outputs = txData.outputs.map((output: BitcoinCash.PsbtTxOutput) => {
+    const outputs = txData.outputs.map((output: any) => {
       const to = output.address ? output.address : txData.from;
       return {
         address: to,