+++ packages/litecoin/src/signers/private-key.signer.ts
@@ -3,6 +3,7 @@ import { Signer, SignerDecorator } from '@xdefi-tech/chains-core';
 import coininfo from 'coininfo';
 import { secp256k1 } from '@noble/curves/secp256k1';
 import * as btc from '@scure/btc-signer';
+import * as Litecoin from 'bitcoinjs-lib';
 
 import { ChainMsg } from '../msg';
 
@@ -62,6 +63,22 @@ export class PrivateKeySigner extends Signer.Provider {
     psbt.finalize();
 
     message.sign(psbt.hex);
+  }
+
+  async signRawTransaction(
+    txHex: string,
+    derivation?: string
+  ): Promise<string> {
+    const network = coininfo.litecoin.main.toBitcoinJS();
+    const pk = Litecoin.ECPair.fromWIF(
+      await this.getPrivateKey(derivation ?? ''),
+      network
+    );
+    const psbt = Litecoin.Psbt.fromHex(txHex, { network });
+
+    psbt.signAllInputs(pk);
+    psbt.finalizeAllInputs();
+    return psbt.extractTransaction(true).toHex();
   }
 }