+++ packages/utxo/src/manifests.ts
@@ -1,6 +1,6 @@
 import { Chain } from '@xdefi-tech/chains-core';
 
 export interface UTXOManifest extends Chain.Manifest {
-  dataProviderType: 'haskoin' | 'blockchair' | 'custom';
-  dataProviderURL: string;
+  dataProviderType?: 'haskoin' | 'blockchair' | 'custom';
+  dataProviderURL?: string;
 }