+++ packages/tron/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @xdefi-tech/chains-tron
 
+## 2.0.14
+
+### Patch Changes
+
+- 62105c4: Feat: update README.md
+
+## 2.0.13
+
+### Patch Changes
+
+- 5b781e6: Feat: add getAccount method to fallbackDataSource in core lib
+- Updated dependencies [5b781e6]
+  - @xdefi-tech/chains-core@2.0.8
+
 ## 2.0.12
 
 ### Patch Changes