+++ packages/utxo/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @xdefi-tech/chains-utxo
 
+## 2.0.7
+
+### Patch Changes
+
+- 7de45e1: Feat: move dataprovider from haskoin & blockchair to indexer for all UTXO chains
+- Updated dependencies [7de45e1]
+  - @xdefi-tech/chains-core@2.0.9
+
+## 2.0.6
+
+### Patch Changes
+
+- a04a730: Feat: add Uint8Array support for memo field
+
+## 2.0.5
+
+### Patch Changes
+
+- 5b781e6: Feat: add getAccount method to fallbackDataSource in core lib
+- Updated dependencies [5b781e6]
+  - @xdefi-tech/chains-core@2.0.8
+
 ## 2.0.4
 
 ### Patch Changes